Extraer feedback visual de mark.js
Octavo paso de la descomposición de mark.js: reloj del header,
setVideoState()/setStatusBar() y los dots de progreso de captura facial
(showCaptureProgress, updateCaptureProgress, hideCaptureProgress,
finishCaptureProgress) se mueven a mark/ui-feedback.js. Todas son
funciones puramente visuales sin estado compartido con el resto del
archivo — el módulo memoiza su propia consulta de captureProgress/
captureDots en el primer uso (no al cargar el módulo, para no depender
del timing exacto de evaluación de módulos vs. parseo del DOM).

Se agrega getCaptureDotCount() para el único punto de mark.js
(startAutoIdentifyDwell) que leía captureDots.length directamente.
Refactor puro, sin cambios de comportamiento.
